[stdlib] Simplify splat constructors for SIMD and make it not `ImplicitlyBoolable`
A new, explicit splatting constructor is added, and the splat logic from
the variadic scalar constructor is removed. This ensures that
constructing from a scalar doesn't get deprioritized during overload
resolution (due to variadic signatures having a lower priority, even if
the variadic is passed only a single value). The existing `Float64`
constructor is removed to remove ambiguity with the new generic scalar
constructor; this also fixes some unintentional implicit conversions
that could degrare floating point arithmetic precision. These changes
had a ripple effect that neccessitated removing the `ImplicitlyBoolable`
trait conformance to avoid accidental conversion from `SIMD` to `SIMD`
through `Bool`

Fixes https://github.com/modularml/mojo/issues/3045.

MOJO_MAX_ORIG_REV_ID: c832f47e16ffa62419a5e42ae959dbbcfe289b1f

+++ mojo/max/test/driver/cuda/test_graph.mojo
@@ -92,7 +92,7 @@ def test_mnist():
             .to_tensor[DType.float32, 2]()
         )
 
-        expected_output = 2.511993e04
+        expected_output = Float32(2.511993e04)
         for i in range(10):
             assert_almost_equal(tensor[0, i], expected_output, atol=0.1)
         assert_equal(tensor.rank, 2)
